# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterDto:
    properties:
      age:
        maximum: 120
        minimum: 1
        type: integer
      email:
        type: string
      gender:
        type: string
      password:
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - age
    - email
    - gender
    - password
    - username
    type: object
  auth.ReturnToken:
    properties:
      exp:
        type: integer
      token:
        type: string
    type: object
  exercise.CreateExerciseDto:
    properties:
      description:
        type: string
      image:
        type: string
      muscle:
        items:
          type: string
        type: array
      name:
        type: string
      type:
        items:
          type: string
        type: array
    required:
    - description
    - image
    - muscle
    - name
    - type
    type: object
  exercise.Exercise:
    properties:
      description:
        type: string
      id:
        type: string
      image:
        type: string
      muscle:
        items:
          type: string
        type: array
      name:
        type: string
      type:
        items:
          type: string
        type: array
    required:
    - description
    - image
    - muscle
    - name
    - type
    type: object
  exercise.UpdateExerciseDto:
    properties:
      description:
        type: string
      image:
        type: string
      muscle:
        items:
          type: string
        type: array
      name:
        type: string
      type:
        items:
          type: string
        type: array
    type: object
...
  user.User:
    properties:
      activitylevel:
        default: 0
        description: default:"0" is not working
        type: integer
      age:
        maximum: 120
        minimum: 1
        type: integer
      created_at:
        default: "null"
        type: string
      email:
        type: string
      gender:
        type: string
      height:
        default: 0
        description: default:"0" is not working
        type: number
      hip:
        default: 0
        description: default:"0" is not working
        type: number
      id:
        type: string
      neck:
        default: 0
        description: default:"0" is not working
        type: number
      password:
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
      waist:
        default: 0
        description: default:"0" is not working
        type: number
      weight:
        default: 0
        description: default:"0" is not working
        type: number
    required:
    - age
    - email
    - gender
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for GymsBro API.
  title: GymsBro API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ReturnToken'
        "400":
          description: Bad Request
          schema: {}
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get me
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema: {}
      summary: Get me
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register
      parameters:
      - description: Register
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema: {}
      summary: Register
      tags:
      - auth
  /exercises:
    get:
      consumes:
      - application/json
      description: Get all exercises
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.Exercise'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get all exercises
      tags:
      - exercises
    post:
      consumes:
      - application/json
      description: Create an exercise
      parameters:
      - description: Create Exercise
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/exercise.CreateExerciseDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/exercise.Exercise'
        "400":
          description: Bad Request
          schema: {}
      summary: Create an exercise
      tags:
      - exercises
  /exercises/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an exercise
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema: {}
        "400":
          description: Bad Request
          schema: {}
      summary: Delete an exercise
      tags:
      - exercises
    get:
      consumes:
      - application/json
      description: Get an exercise
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.Exercise'
        "400":
          description: Bad Request
          schema: {}
      summary: Get an exercise
      tags:
      - exercises
    put:
      consumes:
      - application/json
      description: Update an exercise
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Exercise
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/exercise.UpdateExerciseDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.Exercise'
        "400":
          description: Bad Request
          schema: {}
      summary: Update an exercise
      tags:
      - exercises
  /exercises/many:
    post:
      consumes:
      - application/json
      description: Create many exercises
      parameters:
      - description: Create Exercises
        in: body
        name: exercises
        required: true
        schema:
          items:
            $ref: '#/definitions/exercise.CreateExerciseDto'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/exercise.Exercise'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Create many exercises
      tags:
      - exercises
  /exercises/type/{type}:
    get:
      consumes:
      - application/json
      description: Get exercises by type
      parameters:
      - description: Exercise Type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.Exercise'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get exercises by type
      tags:
      - exercises
...
schemes:
- http
- https
swagger: "2.0"